gun: merge gun object utilities
