+++ src/tr1/game/gun/gun.c
@@ -230,38 +230,6 @@ GAME_OBJECT_ID Gun_GetLaraAnim(const LARA_GUN_TYPE gun_type)
     }
 }
 
-GAME_OBJECT_ID Gun_GetWeaponAnim(const LARA_GUN_TYPE gun_type)
-{
-    switch (gun_type) {
-    case LGT_PISTOLS:
-        return O_LARA_PISTOLS;
-    case LGT_MAGNUMS:
-        return O_LARA_MAGNUMS;
-    case LGT_UZIS:
-        return O_LARA_UZIS;
-    case LGT_SHOTGUN:
-        return O_LARA_SHOTGUN;
-    default:
-        return NO_OBJECT;
-    }
-}
-
-LARA_GUN_TYPE Gun_GetType(const GAME_OBJECT_ID obj_id)
-{
-    switch (obj_id) {
-    case O_PISTOL_ITEM:
-        return LGT_PISTOLS;
-    case O_MAGNUM_ITEM:
-        return LGT_MAGNUMS;
-    case O_UZI_ITEM:
-        return LGT_UZIS;
-    case O_SHOTGUN_ITEM:
-        return LGT_SHOTGUN;
-    default:
-        return LGT_UNARMED;
-    }
-}
-
 void Gun_DrawFlash(LARA_GUN_TYPE weapon_type, int32_t clip)
 {
     int32_t light;
