+++ src/tr2/game/gun/gun.c
@@ -275,111 +275,6 @@ void Gun_InitialiseNewWeapon(void)
     }
 }
 
-int32_t Gun_GetWeaponAnim(const LARA_GUN_TYPE gun_type)
-{
-    switch (gun_type) {
-    case LGT_UNARMED:
-        return O_LARA;
-    case LGT_PISTOLS:
-        return O_LARA_PISTOLS;
-    case LGT_MAGNUMS:
-        return O_LARA_MAGNUMS;
-    case LGT_UZIS:
-        return O_LARA_UZIS;
-    case LGT_SHOTGUN:
-        return O_LARA_SHOTGUN;
-    case LGT_M16:
-        return O_LARA_M16;
-    case LGT_GRENADE:
-        return O_LARA_GRENADE;
-    case LGT_HARPOON:
-        return O_LARA_HARPOON;
-    default:
-        return NO_OBJECT;
-    }
-}
-
-GAME_OBJECT_ID Gun_GetGunObject(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_PISTOLS: return O_PISTOL_ITEM;
-    case LGT_MAGNUMS: return O_MAGNUM_ITEM;
-    case LGT_UZIS:    return O_UZI_ITEM;
-    case LGT_SHOTGUN: return O_SHOTGUN_ITEM;
-    case LGT_HARPOON: return O_HARPOON_ITEM;
-    case LGT_M16:     return O_M16_ITEM;
-    case LGT_GRENADE: return O_GRENADE_ITEM;
-    default:          return NO_OBJECT;
-    }
-    // clang-format on
-}
-
-LARA_GUN_TYPE Gun_GetType(const GAME_OBJECT_ID obj_id)
-{
-    // clang-format off
-    switch (obj_id) {
-    case O_PISTOL_ITEM:  return LGT_PISTOLS;
-    case O_MAGNUM_ITEM:  return LGT_MAGNUMS;
-    case O_UZI_ITEM:     return LGT_UZIS;
-    case O_SHOTGUN_ITEM: return LGT_SHOTGUN;
-    case O_HARPOON_ITEM: return LGT_HARPOON;
-    case O_M16_ITEM:     return LGT_M16;
-    case O_GRENADE_ITEM: return LGT_GRENADE;
-    default:             return LGT_UNARMED;
-    }
-    // clang-format on
-}
-
-GAME_OBJECT_ID Gun_GetAmmoObject(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_PISTOLS: return O_PISTOL_AMMO_ITEM;
-    case LGT_MAGNUMS: return O_MAGNUM_AMMO_ITEM;
-    case LGT_UZIS:    return O_UZI_AMMO_ITEM;
-    case LGT_SHOTGUN: return O_SHOTGUN_AMMO_ITEM;
-    case LGT_HARPOON: return O_HARPOON_AMMO_ITEM;
-    case LGT_M16:     return O_M16_AMMO_ITEM;
-    case LGT_GRENADE: return O_GRENADE_AMMO_ITEM;
-    default:          return NO_OBJECT;
-    }
-    // clang-format on
-}
-
-int32_t Gun_GetAmmoQuantity(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_PISTOLS: return 1;
-    case LGT_MAGNUMS: return MAGNUM_AMMO_QTY;
-    case LGT_UZIS:    return UZI_AMMO_QTY;
-    case LGT_SHOTGUN: return SHOTGUN_AMMO_QTY;
-    case LGT_HARPOON: return HARPOON_AMMO_QTY;
-    case LGT_M16:     return M16_AMMO_QTY;
-    case LGT_GRENADE: return GRENADE_AMMO_QTY;
-    default:          return -1;
-    }
-    // clang-format on
-}
-
-AMMO_INFO *Gun_GetAmmoInfo(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_PISTOLS: return &g_Lara.pistol_ammo;
-    case LGT_MAGNUMS: return &g_Lara.magnum_ammo;
-    case LGT_UZIS:    return &g_Lara.uzi_ammo;
-    case LGT_SHOTGUN: return &g_Lara.shotgun_ammo;
-    case LGT_HARPOON: return &g_Lara.harpoon_ammo;
-    case LGT_M16:     return &g_Lara.m16_ammo;
-    case LGT_GRENADE: return &g_Lara.grenade_ammo;
-    case LGT_SKIDOO:  return &g_Lara.pistol_ammo;
-    default:          return nullptr;
-    }
-    // clang-format on
-}
-
 void Gun_UpdateLaraMeshes(const GAME_OBJECT_ID obj_id)
 {
     const bool lara_has_rifle = Inv_RequestItem(O_SHOTGUN_ITEM)
