+++ src/tr2/game/gun/gun.h
@@ -4,15 +4,6 @@
 
 #include <libtrx/game/gun.h>
 
-int32_t Gun_GetWeaponAnim(LARA_GUN_TYPE gun_type);
-
-// TODO: make this a struct
-LARA_GUN_TYPE Gun_GetType(GAME_OBJECT_ID obj_id);
-GAME_OBJECT_ID Gun_GetGunObject(LARA_GUN_TYPE gun_type);
-GAME_OBJECT_ID Gun_GetAmmoObject(LARA_GUN_TYPE gun_type);
-int32_t Gun_GetAmmoQuantity(LARA_GUN_TYPE gun_type);
-AMMO_INFO *Gun_GetAmmoInfo(LARA_GUN_TYPE gun_type);
-
 void Gun_UpdateLaraMeshes(GAME_OBJECT_ID obj_id);
 void Gun_SetLaraHandLMesh(LARA_GUN_TYPE weapon_type);
 void Gun_SetLaraHandRMesh(LARA_GUN_TYPE weapon_type);
